requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Keras==2.3.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
tensorflow==2.2.0

